+++ integrations/amazon_bedrock/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [integrations/amazon_bedrock-v6.10.0] - 2026-05-12
+
+### 🚀 Features
+
+- Support adaptive_thinking_effort flat param for Claude (#3297)
+
+
+## [integrations/amazon_bedrock-v6.9.0] - 2026-05-11
+
+### 🐛 Bug Fixes
+
+- Normalize token usage conversion in AmazonBedrockGenerator (#3247)
+
+
+## [integrations/amazon_bedrock-v6.8.2] - 2026-05-06
+
+### 🐛 Bug Fixes
+
+- *(amazon-bedrock)* Use dataclasses.replace instead of mutating StreamingChunk (#3273)
+
+
 ## [integrations/amazon_bedrock-v6.8.1] - 2026-04-22
 
 ### 🐛 Bug Fixes